added dotprodsimp to adjugate
